+++ src/main/java/com/bridgelabz/EmployeePayrollApp/service/EmployeePayrollService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface EmployeePayrollService {
 
   List<Employee> getEmployeePayrollData();
-  Employee getEmployeePayrollDataById(int empID);
+  Employee getEmployeePayrollDataById(long empID);
   Employee  createEmployee(EmployeeDTO employeeDTO);
   Employee  updateEmployeePayrollData(int empId,EmployeeDTO employeeDTO);
   void deleteEmployeePayrollData(int empID);
